2-HttpTrigger/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
 
azure-functions
PIL
io
base64